Move "build" from Environment to Processes.
