+++ pylib-v1/ailets/cons/cons.py
@@ -102,71 +102,6 @@ class Environment(IEnvironment):
             explain=node.explain,
         )
 
-    async def build_node_alone(self, nodereg: INodeRegistry, name: str) -> None:
-        """Build a node. Does not build its dependencies."""
-        node = self.get_node(name)
-
-        deps = list(self.iter_deps(name))
-        for dep in deps:
-            dep_name = dep.source
-            if not self.is_node_built(dep_name):
-                raise ValueError(f"Dependency node '{dep_name}' is not built")
-
-        runtime = NodeRuntime(self, nodereg, self._streams, node.name, deps)
-
-        # Execute the node's function with all dependencies
-        try:
-            await node.func(runtime)
-        except Exception:
-            print(f"Error building node '{name}'")
-            print(f"Function: {node.func.__name__}")
-            print("Dependencies:")
-            for dep in node.deps:
-                print(f"  {dep.source} ({dep.stream}) -> {dep.name}")
-            raise
-
-    async def build_target(
-        self,
-        nodereg: INodeRegistry,
-        target: str,
-        one_step: bool = False,
-    ) -> None:
-        """Build nodes in order.
-
-        Args:
-            env: Environment to build in
-            target: Target node to build
-            one_step: If True, build only one step and exit
-        """
-
-        # Get initial plan
-        plan = self.plan(target)
-        current_node_count = len(self.nodes)
-
-        while True:
-            next_node = None
-            for node_name in plan:
-                node = self.get_node(node_name)
-                if not self.is_node_built(node_name):
-                    next_node = node
-                    break
-            # If no dirty nodes, we're done
-            if next_node is None:
-                break
-
-            # Build the node
-            await self.build_node_alone(nodereg, next_node.name)
-
-            # Check if number of nodes changed
-            new_node_count = len(self.nodes)
-            if new_node_count != current_node_count:
-                # Recalculate plan
-                plan = self.plan(target)
-                current_node_count = new_node_count
-
-            if one_step:  # Exit after building one node if requested
-                break
-
     def plan(self, target: str) -> Sequence[str]:
         """Return nodes in build order for the target.
